Authentication flow working

+++ tasks/views.py
@@ -2,15 +2,9 @@ from django.shortcuts import render, redirect, get_object_or_404
 from .models import Task
 from .forms import *
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 
-def get_tasks_to_show(request, tasks): 
-    if request.GET and request.GET.get('filter'):
-        if request.GET.get('filter') == 'complete':
-            return tasks.filter(complete=True)
-        if request.GET.get('filter') == 'incomplete':
-            return tasks.filter(complete=False)
-    return tasks
-
+@login_required
 def index(request):
     tasks = Task.objects.all()
     
@@ -26,6 +20,15 @@ def index(request):
     }
     return render(request, 'tasks/list.html', context)
 
+def get_tasks_to_show(request, tasks): 
+    if request.GET and request.GET.get('filter'):
+        if request.GET.get('filter') == 'complete':
+            return tasks.filter(complete=True)
+        if request.GET.get('filter') == 'incomplete':
+            return tasks.filter(complete=False)
+    return tasks
+
+@login_required
 def add_task(request):
     form = TaskForm()
     
@@ -42,7 +45,8 @@ def add_task(request):
         'title': 'Create Task'    
     }
     return render(request, 'tasks/task_form.html', context)
-    
+
+@login_required
 def updateTask(request, id):
     task = get_object_or_404(Task, pk=id)
     #task = Task.objects.get(id=id)
@@ -64,13 +68,16 @@ def updateTask(request, id):
 
     return render(request, 'tasks/task_form.html', context)
 
-
+@login_required
 def deleteTask(request, id):
     task = get_object_or_404(Task, pk=id)
 
     if request.method == 'POST':
+        if task.owner != request.user:
+            messages.add_message(request, messages.Error, "You can't delete this task because you are not the owner")
+            return redirect('/')
         task.delete()
-        messages.add_message(request, messages.SUCCESS, "Task Created Successfully")
+        messages.add_message(request, messages.SUCCESS, "Task deleted Successfully")
         return redirect('/')
 
     context = {'task': task}
